// src/types/spells.ts
export interface Spell {
  bonus_action: boolean;
  casting_time: string;
  classes: string[];
  components: string;
  concentration: boolean;
  description: string;
  duration: string;
  prepared: boolean;
  level: number;
  name: string;
  range: string;
  ritual: boolean;
  school: string;
  subclasses: string[];
}